refactor: use Repository to create table rth_stripe_php_session
